dispatcher/requirements.txt
psycopg2
openai
python-dotenv
# smtplib
